requirements.txt
numpy~=1.21.5
pandas~=1.4.4
pytorch~=1.13.0
torchvision~=0.14.0
torchaudio~=0.13.0
transformers~=4.24.0
scikit-learn~=1.1.3
scipy~=1.9.3
tqdm~=4.64.0

